hateoas-v

+++ ventas/ventas/src/main/java/com/perfulandia/ventas/repository/VentaRepository.java
@@ -3,5 +3,5 @@ package com.perfulandia.ventas.repository;
 import com.perfulandia.ventas.entity.Venta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VentaRepository extends JpaRepository<Venta, Long> {
+public interface VentaRepository extends JpaRepository<Venta, Integer> {
 }
